fix: change all int to size_t and redefine null

+++ DataStructureVisualizer/include/Model/HeapStructure/MinHeap.h
@@ -5,7 +5,8 @@
 
 class MinHeap : public IHeapStructure {
     protected:
-        bool compare(int idx1, int idx2) override;
+        bool compare(size_t idx1, size_t idx2) override;
+        
     public:
         StructureType getStructureType() const override;
 };
