+++ DataStructureVisualizer/src/Model/HeapStructure/MaxHeap.cpp
@@ -1,6 +1,6 @@
 #include "Model/HeapStructure/MaxHeap.h"
 
-bool MaxHeap::compare(int idx1, int idx2) {
+bool MaxHeap::compare(size_t idx1, size_t idx2) {
     return heapArray[idx1] > heapArray[idx2];
 }
